GH-761 merged three coordinators together
