+++ SafeUIKit/SafeAppUI/AppFlow/MainFlow/MainFlowCoordinator.swift
@@ -7,7 +7,7 @@ import MultisigWalletApplication
 import IdentityAccessApplication
 import Common
 
-final class MainFlowCoordinator: FlowCoordinator {
+open class MainFlowCoordinator: FlowCoordinator {
 
     private let manageTokensFlowCoordinator = ManageTokensFlowCoordinator()
     private let connectExtensionFlowCoordinator = ConnectBrowserExtensionFlowCoordinator()
@@ -24,14 +24,23 @@ final class MainFlowCoordinator: FlowCoordinator {
 
     var transactionSubmissionHandler = TransactionSubmissionHandler()
 
-    override func setUp() {
+    public init() {
+        super.init(rootViewController: UINavigationController())
+        configureGloabalAppearance()
+    }
+
+    open override func setUp() {
         super.setUp()
+        showInitialScreen()
+    }
+
+    func showMainScreen() {
         let mainVC = MainViewController.create(delegate: self)
         mainVC.navigationItem.backBarButtonItem = backButton()
         push(mainVC)
     }
 
-    func receive(message: [AnyHashable: Any]) {
+    open func receive(message: [AnyHashable: Any]) {
         guard let transactionID = walletService.receive(message: message) else { return }
         if let vc = navigationController.topViewController as? ReviewTransactionViewController {
             let tx = ApplicationServiceRegistry.walletService.transactionData(transactionID)!
@@ -53,7 +62,145 @@ final class MainFlowCoordinator: FlowCoordinator {
                                action: nil)
     }
 
+    // FROM AppFlowCoordinator
+
+    private var lockedViewController: UIViewController!
+    private var applicationRootViewController: UIViewController? {
+        get { return UIApplication.shared.keyWindow?.rootViewController }
+        set { UIApplication.shared.keyWindow?.rootViewController = newValue }
+    }
+
+    private var shouldLockWhenAppActive: Bool {
+        return authenticationService.isUserRegistered  && !authenticationService.isUserAuthenticated
+    }
+
+    private func configureGloabalAppearance() {
+        let barButtonAppearance = UIBarButtonItem.appearance(whenContainedInInstancesOf: [UINavigationBar.self])
+        barButtonAppearance.tintColor = ColorName.aquaBlue.color
+
+        let buttonAppearance = UIButton.appearance()
+        buttonAppearance.tintColor = ColorName.aquaBlue.color
+
+        let navBarAppearance = UINavigationBar.appearance()
+        navBarAppearance.barTintColor = .white
+        navBarAppearance.isTranslucent = false
+        navBarAppearance.setBackgroundImage(UIImage(), for: .default)
+        navBarAppearance.shadowImage = Asset.shadow.image
+    }
+
+    func showInitialScreen() {
+        if walletService.hasReadyToUseWallet {
+            showMainScreen()
+        } else {
+            showOnboarding()
+        }
+        lockedViewController = rootViewController
+
+        if authenticationService.isUserRegistered {
+            applicationRootViewController = unlockController { [unowned self] success in
+                guard success else { return }
+                self.applicationRootViewController = self.lockedViewController
+            }
+        } else {
+            applicationRootViewController = lockedViewController
+        }
+    }
+
+    private func unlockController(completion: @escaping (Bool) -> Void) -> UIViewController {
+        return UnlockViewController.create(completion: completion)
+    }
+
+    open func appEntersForeground() {
+        guard let rootVC = applicationRootViewController,
+            !(rootVC is UnlockViewController) && shouldLockWhenAppActive else {
+                return
+        }
+        lockedViewController = rootVC
+        applicationRootViewController = unlockController { [unowned self] success in
+            guard success else { return }
+            self.applicationRootViewController = self.lockedViewController
+        }
+    }
+
+    // iOS: for unknown reason, when alert or activity controller was presented and we
+    // set the UIWindow's root to the root controller that presented that alert,
+    // then all the views (and controllers) under the presented alert are removed when the app
+    // enters foreground.
+    // Dismissing such alerts and controllers after minimizing the app helps.
+    open func appEnteredBackground() {
+        if let presentedVC = applicationRootViewController?.presentedViewController,
+            presentedVC is UIAlertController || presentedVC is UIActivityViewController {
+            presentedVC.dismiss(animated: false, completion: nil)
+        }
+    }
+
+    // FROM onboardingFlowCoordinator
+    let masterPasswordFlowCoordinator = MasterPasswordFlowCoordinator()
+    let setupSafeFlowCoordinator = SetupSafeFlowCoordinator()
+
+    func showOnboarding() {
+        if isUserRegistered {
+            enterSetupSafeFlow()
+        } else {
+            push(StartViewController.create(delegate: self))
+        }
+    }
+
+    private var isUserRegistered: Bool {
+        return ApplicationServiceRegistry.authenticationService.isUserRegistered
+    }
+
+    private func enterSetupSafeFlow() {
+        enter(flow: setupSafeFlowCoordinator) { [unowned self] in
+            self.exitOnboarding()
+        }
+    }
+
+    func  exitOnboarding() {
+        clearNavigationStack()
+        showMainScreen()
+    }
 }
+
+
+
+extension MainFlowCoordinator: StartViewControllerDelegate {
+
+    func didStart() {
+        let controller = TermsAndConditionsViewController.create()
+        controller.delegate = self
+        controller.modalPresentationStyle = .overFullScreen
+        rootViewController.definesPresentationContext = true
+        presentModally(controller)
+    }
+
+}
+
+extension MainFlowCoordinator: TermsAndConditionsViewControllerDelegate {
+
+    public func wantsToOpenTermsOfUse() {
+        SupportFlowCoordinator(from: self).openTermsOfUse()
+    }
+
+    public func wantsToOpenPrivacyPolicy() {
+        SupportFlowCoordinator(from: self).openPrivacyPolicy()
+    }
+
+    public func didDisagree() {
+        dismissModal()
+    }
+
+    public func didAgree() {
+        dismissModal { [unowned self] in
+            self.enter(flow: self.masterPasswordFlowCoordinator) {
+                self.clearNavigationStack()
+                self.enterSetupSafeFlow()
+            }
+        }
+    }
+
+}
+
 
 extension MainFlowCoordinator: MainViewControllerDelegate {
 
@@ -97,7 +244,7 @@ extension MainFlowCoordinator: MainViewControllerDelegate {
 
 extension MainFlowCoordinator: TransactionsTableViewControllerDelegate {
 
-    func didSelectTransaction(id: String) {
+    public func didSelectTransaction(id: String) {
         let controller = TransactionDetailsViewController.create(transactionID: id)
         controller.delegate = self
         push(controller)
@@ -107,7 +254,7 @@ extension MainFlowCoordinator: TransactionsTableViewControllerDelegate {
 
 extension MainFlowCoordinator: TransactionDetailsViewControllerDelegate {
 
-    func showTransactionInExternalApp(from controller: TransactionDetailsViewController) {
+    public func showTransactionInExternalApp(from controller: TransactionDetailsViewController) {
         SupportFlowCoordinator(from: self).openTransactionBrowser(controller.transactionID!)
     }
 
@@ -123,11 +270,11 @@ extension MainFlowCoordinator: SendInputViewControllerDelegate {
 
 extension MainFlowCoordinator: ReviewTransactionViewControllerDelegate {
 
-    func wantsToSubmitTransaction(_ completion: @escaping (Bool) -> Void) {
+    public func wantsToSubmitTransaction(_ completion: @escaping (Bool) -> Void) {
         transactionSubmissionHandler.submitTransaction(from: self, completion: completion)
     }
 
-    func didFinishReview() {
+    public func didFinishReview() {
         let popAction = { [unowned self] in
             self.popToLastCheckpoint()
             self.showTransactionList()
