+++ SafeUIKit/SafeAppUI/AppFlow/MainFlow/OnboardingFlow/SetupSafeFlow/RecoverSafe/RecoverSafeTrackingEvent.swift
