+++ safe/safe/AppDelegate.swift
@@ -22,7 +22,7 @@ import UserNotifications
 class AppDelegate: UIResponder, UIApplicationDelegate, Resettable {
 
     var window: UIWindow?
-    lazy var coordinator = AppFlowCoordinator()
+    lazy var coordinator = MainFlowCoordinator()
     var identityAccessDB: Database?
     var multisigWalletDB: Database?
     var secureStore: SecureStore?
